refactor(bazel): moved FilterableEnumValue complexity for workspace rule property from BazelDetectableOptions to DetectableOptionFactory.
