+++ detectable/build.gradle
@@ -1,6 +1,4 @@
 dependencies {
-    implementation project(':configuration')
-
     implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:2.9.5'
     implementation 'org.freemarker:freemarker:2.3.26-incubating'
     implementation 'com.moandjiezana.toml:toml4j:0.7.1'
